Make the base driver now the default browserstack driver Remove all duplicate capability code
