+++ lib/ca_testing/drivers/v4/remote.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "selenium/webdriver/remote"
-
 require "ca_testing/drivers/v4/options"
 
 module CaTesting
